refactor(desktop): extract the auto-check scheduler into updater-scheduler
Mechanical move of createDesktopUpdaterScheduler, its scheduler/silent
startup option types, and the scheduled-delay floor out of updater.ts.
updater.ts re-exports the scheduler surface so wiring in index.ts and
existing imports stay unchanged; the pure-mock scheduler test cases move
to updater-scheduler.test.ts. Behavior and total test count are
unchanged.
