+++ apps/desktop/src/main/updater.ts
@@ -82,6 +82,10 @@ export {
   type DesktopUpdaterConfig,
   type DesktopUpdaterConfigInput,
 } from "./updater-config.js";
+export {
+  createDesktopUpdaterScheduler,
+  type DesktopUpdaterScheduler,
+} from "./updater-scheduler.js";
 
 const OWNERSHIP_SENTINEL = ".open-design-updater-root.json";
 const STORE_METADATA_FILE = "metadata.json";
@@ -97,7 +101,6 @@ const LOCK_OWNER_FILE = "owner.json";
 const UPDATE_ROOT_VERSION = 1;
 const STORE_METADATA_VERSION = 1;
 const RELEASE_CLEANUP_DESCRIPTOR_VERSION = 1;
-const MIN_SCHEDULED_POLL_DELAY_MS = 1000;
 const DEFERRED_INSTALLER_TIMEOUT_MS = 10 * 60 * 1000;
 const ARTIFACT_DOWNLOAD_MAX_ATTEMPTS = 3;
 const execFileAsync = promisify(execFile);
@@ -124,7 +127,7 @@ export type LauncherPayloadExtractInput = {
   platform: string;
 };
 
-type DesktopUpdaterLogger = Pick<Console, "error" | "warn"> & Partial<Pick<Console, "info">>;
+export type DesktopUpdaterLogger = Pick<Console, "error" | "warn"> & Partial<Pick<Console, "info">>;
 type DetachedProcess = Pick<ReturnType<typeof spawn>, "once" | "unref">;
 type LauncherPayloadCleanupTrigger = "activate" | "manual-clear" | "prepare-existing" | "prepare-promoted";
 type LauncherPayloadCleanupFailure = {
@@ -292,17 +295,6 @@ export type DesktopUpdater = {
   snapshot(): DesktopUpdateStatusSnapshot;
   status(): Promise<DesktopUpdateStatusSnapshot>;
   subscribe(listener: () => void): () => void;
-};
-
-export type DesktopUpdaterScheduler = {
-  isRunning(): boolean;
-  start(): void;
-  stop(reason?: string): void;
-};
-
-type StartupSilentPayloadUpdateOptions = {
-  isEnabled(): Promise<boolean>;
-  requestQuit(): void;
 };
 
 function createError(code: string, message: string, details?: unknown): DesktopUpdateErrorSnapshot {
@@ -3617,137 +3609,5 @@ export function createDesktopUpdater(
         listeners.delete(listener);
       };
     },
-  };
-}
-
-export function createDesktopUpdaterScheduler(
-  updater: DesktopUpdater,
-  options: {
-    backoffInitialMs: number;
-    backoffMaxMs: number;
-    initialDelayMs: number;
-    intervalMs: number;
-    logger?: DesktopUpdaterLogger;
-    startupSilentPayloadUpdate?: StartupSilentPayloadUpdateOptions;
-  },
-): DesktopUpdaterScheduler {
-  const logger = options.logger ?? console;
-  let running = false;
-  let timer: ReturnType<typeof setTimeout> | null = null;
-  let failureCount = 0;
-  let tickRunning = false;
-  let unsubscribe: (() => void) | null = null;
-  let warnedZeroDelay = false;
-  let startupTickPending = true;
-
-  const clearTimer = () => {
-    if (timer == null) return;
-    clearTimeout(timer);
-    timer = null;
-  };
-
-  const stop = (_reason?: string) => {
-    if (!running && timer == null) return;
-    running = false;
-    clearTimer();
-    unsubscribe?.();
-    unsubscribe = null;
-  };
-
-  const normalizeScheduledDelay = (delayMs: number): number => {
-    if (delayMs > 0) return delayMs;
-    if (!warnedZeroDelay) {
-      warnedZeroDelay = true;
-      logger.warn(
-        `[open-design updater] refusing non-positive scheduled poll delay (${delayMs}ms); `
-          + `using ${MIN_SCHEDULED_POLL_DELAY_MS}ms floor`,
-      );
-    }
-    return MIN_SCHEDULED_POLL_DELAY_MS;
-  };
-
-  const nextDelay = (status: DesktopUpdateStatusSnapshot | null): number => {
-    if (status != null && status.state !== DESKTOP_UPDATE_STATES.ERROR && status.error == null) {
-      failureCount = 0;
-      return options.intervalMs;
-    }
-    failureCount += 1;
-    const backoff = options.backoffInitialMs * 2 ** Math.max(0, failureCount - 1);
-    return Math.min(options.backoffMaxMs, backoff);
-  };
-
-  const schedule = (delayMs: number) => {
-    if (!running || timer != null) return;
-    const boundedDelayMs = normalizeScheduledDelay(delayMs);
-    timer = setTimeout(() => {
-      timer = null;
-      void tick();
-    }, boundedDelayMs);
-    timer.unref?.();
-  };
-
-  const tick = async () => {
-    if (!running || tickRunning) return;
-    tickRunning = true;
-    let status: DesktopUpdateStatusSnapshot | null = null;
-    const startupTick = startupTickPending;
-    startupTickPending = false;
-    try {
-      const startupReady = startupTick && options.startupSilentPayloadUpdate != null
-        ? await updater.status()
-        : null;
-      status = await updater.checkForUpdates();
-      if (
-        startupTick
-        && options.startupSilentPayloadUpdate != null
-        && startupReady?.installResult == null
-        && startupReady?.state === DESKTOP_UPDATE_STATES.DOWNLOADED
-        && startupReady.artifact?.type === "payload"
-        && startupReady.capabilities.canApplyInPlace
-        && startupReady.downloadPath != null
-        && startupReady.downloadPath === status.downloadPath
-        && status.installResult == null
-        && status.state === DESKTOP_UPDATE_STATES.DOWNLOADED
-        && status.artifact?.type === "payload"
-        && status.capabilities.canApplyInPlace
-      ) {
-        try {
-          const enabled = await options.startupSilentPayloadUpdate.isEnabled();
-          if (enabled) {
-            status = await updater.installUpdate();
-            if (status.installResult != null) {
-              stop("silent-payload-installed");
-              options.startupSilentPayloadUpdate.requestQuit();
-              return;
-            }
-          }
-        } catch (silentError) {
-          logger.warn("[open-design updater] startup silent payload update failed", silentError);
-        }
-      }
-      if (status.installResult != null) {
-        stop("installer-opened");
-        return;
-      }
-    } catch (error) {
-      logger.warn("[open-design updater] scheduled update check failed", error);
-    } finally {
-      tickRunning = false;
-    }
-    if (running) schedule(nextDelay(status));
-  };
-
-  return {
-    isRunning: () => running,
-    start() {
-      if (running) return;
-      if (updater.snapshot().installResult != null) return;
-      running = true;
-      unsubscribe = updater.subscribe(() => {
-        if (updater.snapshot().installResult != null) stop("installer-opened");
-      });
-      schedule(options.initialDelayMs);
-    },
-    stop,
   };
 }
